Move Config to base util package
