+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/TestOPMode.java
@@ -4,7 +4,7 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Servo;
 
-import org.firstinspires.ftc.teamcode.autonomous.util.Config;
+import org.firstinspires.ftc.teamcode.util.Config;
 
 import java.io.DataOutputStream;
 import java.io.FileOutputStream;
