+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/Config.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomous.util;
+package org.firstinspires.ftc.teamcode.util;
 
 import android.os.Environment;
 import android.util.Log;
